Back-End/requirements.txt
branca==0.6.0
folium==0.14.0
geopandas==0.13.0
keras
matplotlib==3.7.1
numpy
opencv_python
opencv_python_headless
pandas==2.0.2
patchify==0.2.3
Pillow==10.2.0
pylandtemp==0.0.1a1
pyproj==3.5.0
rasterio==1.3.7
scikit_learn==1.2.2
segment_geospatial==0.8.2
segmentation_models==1.0.1
Shapely==2.0.1
split-folders==0.5.1
tensorflow
tqdm==4.65.0
scipy==1.9.3


earthengine-api
geemap
huggingface_hub